move handleSystemPermissionAlertIfNeeded to a separate file
